SQLmodelDay2/pyproject.toml
[tool.poetry]
name = "sqlmodelday2"
version = "0.1.0"
description = ""
authors = ["Muhammad Hashim <hashiimtahir@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
sqlmodel = "^0.0.16"
psycopg = {extras = ["binary"], version = "^3.1.18"}
psycopg2 = "^2.9.9"
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
